labs/lab03-azure-ai-evaluations/scripts/requirements-agent.txt
# Requirements for AI Agent creation script
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
